* improve cert signing integration test+:
+ improve bash utils/test
+ add verify command
+ reorganize certificates commands
+ reorganize certficate configs
+ refactor certificate backup
